remove all jomres syndicates crap
I made a Jomres Plugin to remove ALL this Jomres Syndicates Crap! If the Jomres App is down and you don't have a Jomres License you will not be able to Edit or Add any properties. 
Also, If the Jomres App is down and you don't have a Jomres License and want to edit or add property you will get a cURL error! 

Go to Tools -> Remove Syndicates